Approach works but requires a Buffer Size for the BlockingQueue almost as large as the inbound stream of Primes.  As such, the solution will work up to about 10^9 for Set K=0, but beyond that will be an issue due to java.  The max buffer for BlockingQueue is Integer.MAX_VALUE which won't work beyond 10^9

+++ src/com/primefractal/main/PropertiesHelper.java
@@ -124,9 +124,11 @@ final public class PropertiesHelper extends Properties {
 	
 	
 	public final static int		PRIMES_Q_BUF_SIZE_	=1000000;	// 10^6
+	//public final static int		PRIMES_Q_BUF_SIZE_	=10000000;	// 10^7
+	//public final static int		PRIMES_Q_BUF_SIZE_	=3000000;	// 3M
 	public final static int		HIGH_Q_BUF_SIZE_	=1000000;	// 10^6
-	public final static int		Q_OFFER_SLEEP_DURATION_ = 50; // millis
-	public final static int		Q_POLL_SLEEP_DURATION_ 	= 50; // millis
+	public final static long		Q_OFFER_SLEEP_DURATION_ = 50L; // millis
+	public final static long		Q_POLL_SLEEP_DURATION_ 	= 50L; // millis
 	public final static Long 	EOF_FOR_QUEUE_=new Long(-1);
 	
 	final private static Logger LOGGER_=Logger.getLogger("PrimeTransformationStream");  // project wide logger
